+++ src/descriptor_test.rs
@@ -305,8 +305,8 @@ fn load_not_found() {
 }
 
 #[test]
-fn load_default_no_stable() {
-    let config = load_default(false, false);
+fn load_internal_descriptors_no_stable() {
+    let config = load_internal_descriptors(false, false);
 
     let mut task = config.tasks.get("empty");
     assert!(task.is_some());
@@ -315,8 +315,8 @@ fn load_default_no_stable() {
 }
 
 #[test]
-fn load_default_with_stable() {
-    let config = load_default(true, false);
+fn load_internal_descriptors_with_stable() {
+    let config = load_internal_descriptors(true, false);
 
     let mut task = config.tasks.get("empty");
     assert!(task.is_some());
@@ -325,8 +325,8 @@ fn load_default_with_stable() {
 }
 
 #[test]
-fn load_default_no_experimental() {
-    let config = load_default(true, false);
+fn load_internal_descriptors_no_experimental() {
+    let config = load_internal_descriptors(true, false);
 
     let mut task = config.tasks.get("ci-flow");
     assert!(task.is_some());
@@ -335,8 +335,8 @@ fn load_default_no_experimental() {
 }
 
 #[test]
-fn load_default_with_experimental() {
-    let config = load_default(true, true);
+fn load_internal_descriptors_with_experimental() {
+    let config = load_internal_descriptors(true, true);
 
     let mut task = config.tasks.get("ci-flow");
     assert!(task.is_some());
@@ -479,95 +479,4 @@ fn run_load_script_invalid_load_script() {
     external_config.config = Some(config);
 
     run_load_script(&external_config);
-}
-
-#[test]
-fn list_steps_empty() {
-    let config_section = ConfigSection::new();
-    let env = IndexMap::<String, EnvValue>::new();
-    let tasks = IndexMap::<String, Task>::new();
-    let config = Config {
-        config: config_section,
-        env,
-        tasks,
-    };
-
-    let count = list_steps(&config);
-
-    assert_eq!(count, 0);
-}
-
-#[test]
-fn list_steps_all_public() {
-    let config_section = ConfigSection::new();
-    let env = IndexMap::<String, EnvValue>::new();
-
-    let mut tasks = IndexMap::<String, Task>::new();
-    let mut task1 = Task::new();
-    task1.description = Some("1".to_string());
-    tasks.insert("1".to_string(), task1);
-    let mut task2 = Task::new();
-    task2.description = Some("2".to_string());
-    tasks.insert("2".to_string(), task2);
-
-    let config = Config {
-        config: config_section,
-        env,
-        tasks,
-    };
-
-    let count = list_steps(&config);
-
-    assert_eq!(count, 2);
-}
-
-#[test]
-fn list_steps_all_private() {
-    let config_section = ConfigSection::new();
-    let env = IndexMap::<String, EnvValue>::new();
-
-    let mut tasks = IndexMap::<String, Task>::new();
-    let mut task1 = Task::new();
-    task1.description = Some("1".to_string());
-    task1.private = Some(true);
-    tasks.insert("1".to_string(), task1);
-    let mut task2 = Task::new();
-    task2.description = Some("2".to_string());
-    task2.private = Some(true);
-    tasks.insert("2".to_string(), task2);
-
-    let config = Config {
-        config: config_section,
-        env,
-        tasks,
-    };
-
-    let count = list_steps(&config);
-
-    assert_eq!(count, 0);
-}
-
-#[test]
-fn list_steps_mixed() {
-    let config_section = ConfigSection::new();
-    let env = IndexMap::<String, EnvValue>::new();
-
-    let mut tasks = IndexMap::<String, Task>::new();
-    let mut task1 = Task::new();
-    task1.description = Some("1".to_string());
-    task1.private = Some(true);
-    tasks.insert("1".to_string(), task1);
-    let mut task2 = Task::new();
-    task2.description = Some("2".to_string());
-    tasks.insert("2".to_string(), task2);
-
-    let config = Config {
-        config: config_section,
-        env,
-        tasks,
-    };
-
-    let count = list_steps(&config);
-
-    assert_eq!(count, 1);
 }