+++ src/lib.rs
@@ -133,12 +133,15 @@ mod command;
 mod condition;
 mod config;
 mod descriptor;
+mod diff_steps;
 mod environment;
 mod execution_plan;
 mod installer;
+mod io;
 mod legacy;
+mod list_steps;
 mod logger;
-mod print;
+mod print_steps;
 mod runner;
 mod scriptengine;
 mod storage;
@@ -218,14 +221,18 @@ fn run(cli_args: CliArgs, global_config: &GlobalConfig) {
         None => env_cli_entries,
     };
 
-    let config = descriptor::load(&build_file, force_makefile, env, cli_args.experimental);
+    let experimental = cli_args.experimental;
+    let config = descriptor::load(&build_file, force_makefile, env, experimental);
 
     let env_info = environment::setup_env(&cli_args, &config, &task);
 
     if cli_args.list_all_steps {
-        descriptor::list_steps(&config);
+        list_steps::run(&config, &cli_args.output_format);
+    } else if cli_args.diff_execution_plan {
+        let default_config = descriptor::load_internal_descriptors(true, experimental);
+        diff_steps::run(&default_config, &config, &task, &cli_args);
     } else if cli_args.print_only {
-        print::print(
+        print_steps::print(
             &config,
             &task,
             &cli_args.output_format,
@@ -308,6 +315,7 @@ fn run_for_args(
     cli_args.disable_workspace = cmd_matches.is_present("no-workspace");
     cli_args.disable_on_error = cmd_matches.is_present("no-on-error");
     cli_args.list_all_steps = cmd_matches.is_present("list-steps");
+    cli_args.diff_execution_plan = cmd_matches.is_present("diff-steps");
 
     let default_task_name = match global_config.default_task_name {
         Some(ref value) => value.as_str().clone(),
@@ -426,8 +434,8 @@ fn create_cli<'a, 'b>(
                 .help("Disables the update check during startup"),
         )
         .arg(
-            Arg::from_usage("--output-format=[OUTPUT FORMAT] 'The print steps format'")
-                .possible_values(&["default", "short-description"])
+            Arg::from_usage("--output-format=[OUTPUT FORMAT] 'The print/list steps format (some operations do not support all formats)'")
+                .possible_values(&["default", "short-description", "markdown"])
                 .default_value(DEFAULT_OUTPUT_FORMAT),
         )
         .arg(Arg::with_name("print-steps").long("--print-steps").help(
@@ -438,6 +446,11 @@ fn create_cli<'a, 'b>(
             Arg::with_name("list-steps")
                 .long("--list-all-steps")
                 .help("Lists all known steps"),
+        )
+        .arg(
+            Arg::with_name("diff-steps")
+                .long("--diff-steps")
+                .help("Runs diff between custom flow and prebuilt flow (requires git)"),
         )
         .arg(Arg::with_name("TASK").help("The task name to execute"))
         .arg(