+++ src/scriptengine/generic_script.rs
@@ -8,7 +8,8 @@
 mod generic_script_test;
 
 use crate::command;
-use crate::scriptengine::script_utils::{create_script_file, delete_file};
+use crate::io::delete_file;
+use crate::scriptengine::script_utils::create_script_file;
 
 fn run_file(file: &str, runner: &String) -> bool {
     let exit_code = command::run_command(runner, &Some(vec![file.to_string()]), false);
@@ -17,14 +18,19 @@ fn run_file(file: &str, runner: &String) -> bool {
     exit_code == 0
 }
 
-pub(crate) fn execute(script_text: &Vec<String>, runner: String, extension: String) {
+pub(crate) fn execute(
+    script_text: &Vec<String>,
+    runner: String,
+    extension: String,
+    validate: bool,
+) {
     let file = create_script_file(script_text, &extension);
 
     let valid = run_file(&file, &runner);
 
     delete_file(&file);
 
-    if !valid {
+    if validate && !valid {
         error!("Unable to execute generic script.");
     }
 }