+++ src/scriptengine/generic_script_test.rs
@@ -7,6 +7,7 @@ fn execute_shell() {
         &vec!["exit 0".to_string()],
         test::get_os_runner(),
         test::get_os_extension(),
+        true,
     );
 }
 
@@ -17,5 +18,16 @@ fn execute_shell_error() {
         &vec!["exit 1".to_string()],
         test::get_os_runner(),
         test::get_os_extension(),
+        true,
+    );
+}
+
+#[test]
+fn execute_shell_error_no_validate() {
+    execute(
+        &vec!["exit 1".to_string()],
+        test::get_os_runner(),
+        test::get_os_extension(),
+        false,
     );
 }