+++ src/scriptengine/mod_test.rs
@@ -2,13 +2,22 @@ use super::*;
 use crate::test;
 
 #[test]
+fn get_engine_type_no_runner_no_script() {
+    let task = Task::new();
+
+    let output = get_engine_type(&task);
+
+    assert_eq!(output, EngineType::Unsupported);
+}
+
+#[test]
 fn get_engine_type_no_runner() {
     let mut task = Task::new();
     task.script = Some(vec!["test".to_string()]);
 
     let output = get_engine_type(&task);
 
-    assert_eq!(output, EngineType::Unsupported);
+    assert_eq!(output, EngineType::OS);
 }
 
 #[test]
@@ -69,7 +78,16 @@ fn get_engine_type_generic() {
 #[test]
 fn invoke_no_runner() {
     let mut task = Task::new();
-    task.script = Some(vec!["test".to_string()]);
+    task.script = Some(vec!["echo test".to_string()]);
+
+    let output = invoke(&task, &vec![]);
+
+    assert!(output);
+}
+
+#[test]
+fn invoke_no_script_no_runner() {
+    let task = Task::new();
 
     let output = invoke(&task, &vec![]);