+++ src/scriptengine/rsscript_test.rs
@@ -7,6 +7,7 @@ fn execute_valid() {
         execute(
             &vec!["fn main() {println!(\"test\");}".to_string()],
             &vec![],
+            true,
         );
     }
 }
@@ -15,7 +16,11 @@ fn execute_valid() {
 #[should_panic]
 fn execute_not_compile() {
     if test::should_test(true) {
-        execute(&vec!["fn main() {donotcompile();}".to_string()], &vec![]);
+        execute(
+            &vec!["fn main() {donotcompile();}".to_string()],
+            &vec![],
+            true,
+        );
     }
 }
 
@@ -23,6 +28,21 @@ fn execute_not_compile() {
 #[should_panic]
 fn execute_runtime_panic() {
     if test::should_test(true) {
-        execute(&vec!["fn main() {panic!(\"error\");}".to_string()], &vec![]);
+        execute(
+            &vec!["fn main() {panic!(\"error\");}".to_string()],
+            &vec![],
+            true,
+        );
+    }
+}
+
+#[test]
+fn execute_runtime_panic_no_validate() {
+    if test::should_test(false) {
+        execute(
+            &vec!["fn main() {panic!(\"error\");}".to_string()],
+            &vec![],
+            false,
+        );
     }
 }