+++ src/types_test.rs
@@ -16,6 +16,7 @@ fn cli_args_new() {
     assert!(!cli_args.disable_check_for_updates);
     assert!(!cli_args.print_only);
     assert!(!cli_args.list_all_steps);
+    assert!(!cli_args.diff_execution_plan);
     assert!(!cli_args.experimental);
     assert!(cli_args.arguments.is_none());
     assert_eq!(cli_args.output_format, "default");
@@ -368,6 +369,7 @@ fn task_new() {
     assert!(task.command.is_none());
     assert!(task.disabled.is_none());
     assert!(task.private.is_none());
+    assert!(task.watch.is_none());
     assert!(task.condition.is_none());
     assert!(task.condition_script.is_none());
     assert!(task.description.is_none());
@@ -430,6 +432,7 @@ fn task_extend_both_have_misc_data() {
     base.command = Some("test1".to_string());
     base.disabled = Some(false);
     base.private = Some(false);
+    base.watch = Some(false);
     base.script = Some(vec!["1".to_string(), "2".to_string()]);
 
     let extended = Task {
@@ -441,6 +444,7 @@ fn task_extend_both_have_misc_data() {
         workspace: None,
         disabled: Some(true),
         private: Some(true),
+        watch: Some(true),
         condition: None,
         condition_script: None,
         force: Some(true),
@@ -474,6 +478,7 @@ fn task_extend_both_have_misc_data() {
     assert!(base.workspace.is_none());
     assert!(base.disabled.is_some());
     assert!(base.private.is_some());
+    assert!(base.watch.is_some());
     assert!(base.condition.is_none());
     assert!(base.condition_script.is_none());
     assert!(base.force.is_some());
@@ -503,6 +508,7 @@ fn task_extend_both_have_misc_data() {
     assert_eq!(base.command.unwrap(), "test1");
     assert!(base.disabled.unwrap());
     assert!(base.private.unwrap());
+    assert!(base.watch.unwrap());
     assert!(base.force.unwrap());
     assert_eq!(base.env.unwrap().len(), 0);
     assert_eq!(base.alias.unwrap(), "alias2");
@@ -520,6 +526,7 @@ fn task_extend_extended_have_all_fields() {
         workspace: None,
         disabled: Some(false),
         private: Some(true),
+        watch: Some(true),
         condition: None,
         condition_script: None,
         force: Some(true),
@@ -555,6 +562,7 @@ fn task_extend_extended_have_all_fields() {
         workspace: Some(true),
         disabled: Some(true),
         private: Some(false),
+        watch: Some(false),
         condition: Some(TaskCondition {
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
             channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -586,6 +594,7 @@ fn task_extend_extended_have_all_fields() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            watch: Some(false),
             condition: Some(TaskCondition {
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
                 channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -614,6 +623,7 @@ fn task_extend_extended_have_all_fields() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            watch: Some(false),
             condition: Some(TaskCondition {
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
                 channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -642,6 +652,7 @@ fn task_extend_extended_have_all_fields() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            watch: Some(false),
             condition: Some(TaskCondition {
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
                 channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -676,6 +687,7 @@ fn task_extend_extended_have_all_fields() {
     assert!(base.workspace.is_some());
     assert!(base.disabled.is_some());
     assert!(base.private.is_some());
+    assert!(base.watch.is_some());
     assert!(base.condition.is_some());
     assert!(base.condition_script.is_some());
     assert!(base.force.is_some());
@@ -708,6 +720,7 @@ fn task_extend_extended_have_all_fields() {
     assert!(base.workspace.unwrap());
     assert!(base.disabled.unwrap());
     assert!(!base.private.unwrap());
+    assert!(!base.watch.unwrap());
     assert_eq!(base.condition_script.unwrap().len(), 1);
     assert!(!base.force.unwrap());
     assert_eq!(base.env.unwrap().len(), 1);
@@ -746,6 +759,7 @@ fn task_extend_clear_with_no_data() {
         workspace: Some(false),
         disabled: Some(true),
         private: Some(false),
+        watch: Some(false),
         condition: Some(TaskCondition {
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
             channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -777,6 +791,7 @@ fn task_extend_clear_with_no_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            watch: Some(false),
             condition: Some(TaskCondition {
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
                 channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -805,6 +820,7 @@ fn task_extend_clear_with_no_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            watch: Some(false),
             condition: Some(TaskCondition {
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
                 channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -833,6 +849,7 @@ fn task_extend_clear_with_no_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            watch: Some(false),
             condition: Some(TaskCondition {
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
                 channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -869,6 +886,7 @@ fn task_extend_clear_with_no_data() {
     assert!(base.workspace.is_none());
     assert!(base.disabled.is_none());
     assert!(base.private.is_none());
+    assert!(base.watch.is_none());
     assert!(base.condition.is_none());
     assert!(base.condition_script.is_none());
     assert!(base.force.is_none());
@@ -907,6 +925,7 @@ fn task_extend_clear_with_all_data() {
         workspace: Some(true),
         disabled: Some(true),
         private: Some(false),
+        watch: Some(false),
         condition: Some(TaskCondition {
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
             channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -938,6 +957,7 @@ fn task_extend_clear_with_all_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            watch: Some(false),
             condition: Some(TaskCondition {
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
                 channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -966,6 +986,7 @@ fn task_extend_clear_with_all_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            watch: Some(false),
             condition: Some(TaskCondition {
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
                 channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -994,6 +1015,7 @@ fn task_extend_clear_with_all_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            watch: Some(false),
             condition: Some(TaskCondition {
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
                 channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -1027,6 +1049,7 @@ fn task_extend_clear_with_all_data() {
     assert!(base.workspace.is_some());
     assert!(base.disabled.is_some());
     assert!(base.private.is_some());
+    assert!(base.watch.is_some());
     assert!(base.condition.is_some());
     assert!(base.condition_script.is_some());
     assert!(base.force.is_some());
@@ -1098,6 +1121,7 @@ fn task_get_normalized_task_undefined() {
         command: Some("command".to_string()),
         disabled: Some(false),
         private: Some(true),
+        watch: Some(true),
         condition: None,
         condition_script: None,
         force: None,
@@ -1127,6 +1151,7 @@ fn task_get_normalized_task_undefined() {
     assert!(normalized_task.command.is_some());
     assert!(normalized_task.disabled.is_some());
     assert!(normalized_task.private.is_some());
+    assert!(normalized_task.watch.is_some());
     assert!(normalized_task.condition.is_none());
     assert!(normalized_task.condition_script.is_none());
     assert!(normalized_task.force.is_none());
@@ -1161,6 +1186,7 @@ fn task_get_normalized_task_undefined() {
     assert!(!normalized_task.workspace.unwrap());
     assert!(!normalized_task.disabled.unwrap());
     assert!(normalized_task.private.unwrap());
+    assert!(normalized_task.watch.unwrap());
     assert!(!normalized_task.force.unwrap_or(false));
     assert_eq!(normalized_task.alias.unwrap(), "alias");
     assert_eq!(normalized_task.linux_alias.unwrap(), "linux");
@@ -1196,6 +1222,7 @@ fn task_get_normalized_task_with_override_no_clear() {
         workspace: Some(true),
         disabled: Some(false),
         private: Some(true),
+        watch: Some(true),
         condition: Some(TaskCondition {
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
             channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -1223,6 +1250,7 @@ fn task_get_normalized_task_with_override_no_clear() {
             command: Some("linux_command".to_string()),
             disabled: Some(true),
             private: Some(false),
+            watch: Some(false),
             condition: Some(TaskCondition {
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
                 channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -1264,6 +1292,7 @@ fn task_get_normalized_task_with_override_no_clear() {
     assert!(normalized_task.workspace.is_some());
     assert!(normalized_task.disabled.is_some());
     assert!(normalized_task.private.is_some());
+    assert!(normalized_task.watch.is_some());
     assert!(normalized_task.condition.is_some());
     assert!(normalized_task.condition_script.is_some());
     assert!(normalized_task.force.is_some());
@@ -1296,6 +1325,7 @@ fn task_get_normalized_task_with_override_no_clear() {
     assert!(normalized_task.workspace.unwrap());
     assert!(normalized_task.disabled.unwrap());
     assert!(!normalized_task.private.unwrap());
+    assert!(!normalized_task.watch.unwrap());
     assert_eq!(normalized_task.condition_script.unwrap().len(), 1);
     assert!(normalized_task.force.unwrap());
     assert_eq!(normalized_task.env.unwrap().len(), 1);
@@ -1334,6 +1364,7 @@ fn task_get_normalized_task_with_override_clear_false() {
         workspace: Some(true),
         disabled: Some(false),
         private: Some(true),
+        watch: Some(true),
         condition: Some(TaskCondition {
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
             channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -1360,6 +1391,7 @@ fn task_get_normalized_task_with_override_clear_false() {
             command: Some("linux_command".to_string()),
             disabled: Some(true),
             private: Some(false),
+            watch: Some(false),
             condition: Some(TaskCondition {
                 platforms: Some(vec!["linux".to_string()]),
                 channels: Some(vec![
@@ -1405,6 +1437,7 @@ fn task_get_normalized_task_with_override_clear_false() {
     assert!(normalized_task.workspace.is_some());
     assert!(normalized_task.disabled.is_some());
     assert!(normalized_task.private.is_some());
+    assert!(normalized_task.watch.is_some());
     assert!(normalized_task.condition.is_some());
     assert!(normalized_task.condition_script.is_some());
     assert!(normalized_task.force.is_some());
@@ -1437,6 +1470,7 @@ fn task_get_normalized_task_with_override_clear_false() {
     assert!(normalized_task.workspace.unwrap());
     assert!(normalized_task.disabled.unwrap());
     assert!(!normalized_task.private.unwrap());
+    assert!(!normalized_task.watch.unwrap());
     assert_eq!(normalized_task.condition_script.unwrap().len(), 2);
     assert!(normalized_task.force.unwrap());
     assert_eq!(normalized_task.env.unwrap().len(), 1);
@@ -1470,6 +1504,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
         command: Some("command".to_string()),
         disabled: Some(false),
         private: Some(true),
+        watch: Some(true),
         condition: Some(TaskCondition {
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
             channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -1500,6 +1535,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
             command: None,
             disabled: None,
             private: None,
+            watch: None,
             condition: None,
             condition_script: None,
             force: None,
@@ -1526,6 +1562,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
     assert!(normalized_task.command.is_some());
     assert!(normalized_task.disabled.is_some());
     assert!(normalized_task.private.is_some());
+    assert!(normalized_task.watch.is_some());
     assert!(normalized_task.condition.is_some());
     assert!(normalized_task.condition_script.is_some());
     assert!(normalized_task.force.is_some());
@@ -1557,6 +1594,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
     assert_eq!(normalized_task.command.unwrap(), "command");
     assert!(!normalized_task.disabled.unwrap());
     assert!(normalized_task.private.unwrap());
+    assert!(normalized_task.watch.unwrap());
     assert!(!normalized_task.force.unwrap());
     assert_eq!(normalized_task.env.unwrap().len(), 0);
     assert_eq!(normalized_task.cwd.unwrap(), "cwd".to_string());
@@ -1585,6 +1623,7 @@ fn task_get_normalized_task_with_override_clear_true() {
         command: Some("command".to_string()),
         disabled: Some(false),
         private: Some(true),
+        watch: Some(true),
         condition: Some(TaskCondition {
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
             channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -1615,6 +1654,7 @@ fn task_get_normalized_task_with_override_clear_true() {
             command: None,
             disabled: None,
             private: None,
+            watch: None,
             condition: None,
             condition_script: None,
             force: None,
@@ -1641,6 +1681,7 @@ fn task_get_normalized_task_with_override_clear_true() {
     assert!(normalized_task.command.is_none());
     assert!(normalized_task.disabled.is_none());
     assert!(normalized_task.private.is_none());
+    assert!(normalized_task.watch.is_none());
     assert!(normalized_task.condition.is_none());
     assert!(normalized_task.condition_script.is_none());
     assert!(normalized_task.force.is_none());