+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## CHANGELOG
 
+### v0.16.1
+
+* Enable coverage for every Linux based CI system
+* Support task watch #118
+* Command --list-all-steps now supports markdown output format
+* New command --diff-steps to enable diff changes to flow based on custom makefile #121"
+* Document Azure Pipelines config #161
+* New diff-files task which enables to diff 2 provided files
+
 ### v0.16.0 (2019-01-04)
 
 * Should abort on non existent makefile if provided on cli or on extends property #143 (**backward compatability break**)